[BugFix] Fix Clang compiling Error and vast warning. compile info: http://jenkins-domain/job/opensource_centos8_build_clang/137/consoleFull

+++ sql/item.h
@@ -4613,12 +4613,12 @@ class Item_result_field : public Item {
   }
 
   longlong llrint_with_overflow_check(double realval) {
-    if (realval < LLONG_MIN || realval > LLONG_MAX) {
+    if ((long long)realval < LLONG_MIN || (long long)realval > LLONG_MAX) {
       raise_integer_overflow();
       return error_int();
     }
     // Rounding error, llrint() may return LLONG_MIN.
-    const longlong retval = realval == LLONG_MAX ? LLONG_MAX : llrint(realval);
+    const longlong retval = (long long)realval == LLONG_MAX ? LLONG_MAX : llrint(realval);
     return retval;
   }
